# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Geld.Network Server.",
        "title": "Geld.Network API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/stage/remove/{_id}": {
            "delete": {
                "description": "Delete an existing Stage By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stages"
                ],
                "summary": "Delete Stage By ID",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stage/save": {
            "post": {
                "description": "Creates Stages with Fields unique to it for data input labels by Geld App",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stages"
                ],
                "summary": "Create and Save Stages for Workflows",
                "parameters": [
                    {
                        "description": "Stage Details",
                        "name": "stagesBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Stages"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stage/{_id}": {
            "get": {
                "description": "Get existing Stage By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stages"
                ],
                "summary": "Get Stage By ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/updatestage": {
            "put": {
                "description": "Update an exisiting Stages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stages"
                ],
                "summary": "Update Stage Data",
                "parameters": [
                    {
                        "description": "Stage Details",
                        "name": "stagesBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestDtos.UpdateStages"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/userstage/{userid}": {
            "get": {
                "description": "Retrieves paginated Stage data associated with a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stages"
                ],
                "summary": "Get Paginated Stage Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default from env: PAGINATION_DEFUALT_LIMIT)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Requested page (default from env: PAGINATION_DEFAULT_PAGE)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sort order (-1: Desc, 1: Asc, default from env: PAGINATION_DEFAULT_SORT)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Stages": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stagename": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
        "requestDtos.UpdateStages": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stagename": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
        "responseDtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "responseDtos.ResultResponse": {
            "type": "object",
            "properties": {
                "Response": {},
                "Status": {
                    "type": "integer"
                }
            }
        }
    }
}